+++ react/src/GridSystem.tsx
@@ -275,6 +275,25 @@ function GridSystem(props : props, ref : any) {
   
   return (
     <>
+    <div className="modal" id="myModal">
+        <div className="modal-dialog">
+            <div className="modal-content">
+
+            <div className="modal-header">
+                <h4 className="modal-title">Modal Heading</h4>
+                <button type="button" className="btn-close" data-bs-dismiss="modal"></button>
+            </div>
+            <div className="modal-body">
+                Modal body..
+            </div>
+
+            <div className="modal-footer">
+                <button type="button" className="btn btn-danger" data-bs-dismiss="modal">Close</button>
+            </div>
+
+            </div>
+        </div>
+    </div>
     <GridLayout
       className="layout"
       layout={grid}
@@ -285,6 +304,7 @@ function GridSystem(props : props, ref : any) {
 
       // UpdateGrid only runs when user interacts with grid
       onDragStart={() => {
+        //e.preventDefault();
         userInteracting.current = true;
       }}
       onResizeStart={() => userInteracting.current = true}